LayoutTests/http/tests/websocket/tests/hybi/extensions.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" href="../../../../js-test-resources/js-test-style.css">
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description("Test WebSocket.extensions attribute.");

if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var ws = new WebSocket("ws://localhost:8880/websocket/tests/hybi/simple");
shouldBeEqualToString("ws.extensions", "");

// extensions attribute is read-only.
ws.extensions = "foo";
shouldBeEqualToString("ws.extensions", "");

var successfullyParsed = true;
</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
