LayoutTests/http/tests/websocket/tests/hybi/extensions-expected.txt

Test WebSocket.extensions attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.extensions is ""
PASS ws.extensions is ""
PASS successfullyParsed is true

TEST COMPLETE
